chore: moved nodes around for better reusage
