+++ internal/ast/css/calculations/CSSCalcNumberProduct.ts
