+++ internal/ast/css/calculations/CSSCalcProduct.ts
