+++ internal/css-parser/types.ts
@@ -9,7 +9,7 @@ import {
 } from "@internal/parser-core";
 import {
 	CSSBlock,
-	CSSCalcFunction,
+	CSSCalcFunction, CSSCalcSum,
 	CSSComma,
 	CSSCustomProperty,
 	CSSDimension,
@@ -92,6 +92,7 @@ export type AnyCSSValue =
 	| CSSCustomProperty
 	| CSSUrlFunction
 	| CSSCalcFunction
+	| CSSCalcSum
 	| CSSRaw;
 
 export type AnyFunction = CSSFunction | CSSVarFunction | CSSUrlFunction;
