+++ internal/formatter/builders/css/calculations/CSSCalcNumberProduct.ts
