+++ internal/formatter/builders/css/calculations/CSSCalcProduct.ts
