+++ internal/formatter/test-fixtures/css/function/input.css
@@ -1,7 +1,9 @@
 
 .func {
 	background-color: rgb(0 128 255);
-	border: var(--foo,red, blue, calc( 10px   +1rem));
-
+	border: var(--foo,red, blue, min( 10px   +1rem));
+	padding: calc(10px    - 2rem);
 	background: url("   something ");
+	width: calc(100vw    + 2em    /(2 - 3) / (   2       - 6));
+	height: calc(1px - 2px + 3px);
 }
