+++ internal/formatter/test-fixtures/css/function/input.test.md
@@ -16,16 +16,12 @@
 ```css
 
 .func {
-	padding-left: calc();
-	padding-right: calc( 1px );
-	padding-bottom: calc(1px  - 2px);
-	padding-top: calc(1px, 2px);
-	padding: calc(calc(1px, 2px), 3px);
-	height: calc(1px, 2px, 3px);
 	background-color: rgb(0 128 255);
-	border: var(--foo,red, blue, calc( 10px   +1rem));
-
+	border: var(--foo,red, blue, min( 10px   +1rem));
+	padding: calc(10px    - 2rem);
 	background: url("   something ");
+	width: calc(100vw    + 2em    /(2 - 3) / (   2       - 6));
+	height: calc(1px - 2px + 3px);
 }
 
 ```
@@ -34,16 +30,12 @@
 
 ```css
 .func {
-	padding-left: calc();
-	padding-right: calc(1px);
-	padding-bottom: calc(1px - 2px);
-	padding-top: calc(1px, 2px);
-	padding: calc(calc(1px, 2px), 3px);
-	height: calc(1px, 2px, 3px);
 	background-color: rgb(0 128 255);
-	border: var(--foo, red, blue, calc(10px 1rem));
-
+	border: var(--foo, red, blue, min(10px 1rem));
+	padding: calc(10px - 2rem);
 	background: url("   something ");
+	width: calc(100vw + 2em / (2 - 3) / (2 - 6));
+	height: calc(1px - 2px + 3px);
 }
 
 ```
